cleanup deps
