+++ src/components/layout/CdnTime.tsx
@@ -1,5 +1,16 @@
 import { TwClass, TimeStyle } from "@/deps/styles.ts";
-import { formatTime } from "@/deps/time.ts";
+
+export function formatTime(mtime: Date | null) {
+  if (mtime !== null) {
+    return mtime.toLocaleDateString("en-us", {
+      year: "numeric",
+      month: "long",
+      day: "numeric",
+    });
+  } else {
+    return null;
+  }
+}
 
 export default function getTime({
   inline,
