+++ src/components/layout/Head.tsx
@@ -1,4 +1,5 @@
 import { Head } from "$fresh/src/runtime/head.ts";
+import { CDN_URL } from "@/deps/paths.ts";
 
 export default function HeadComponent() {
   return (
@@ -9,7 +10,7 @@ export default function HeadComponent() {
       <link rel="stylesheet" href="/style/toc.css" />
       <link rel="stylesheet" href="/style/global.css" />
       <link rel="stylesheet" href="/style/code.css" />
-      <link href="https://cdn.cubething.dev/font/Fonts.css" rel="stylesheet" />
+      <link href={CDN_URL + "/font/Fonts.css"} rel="stylesheet" />
       {/*<!-- Starry-night code highlighter theme -->*/}
       <link
         rel="stylesheet"
