+++ src/deps/code-highlight.ts
@@ -1,8 +1,8 @@
 import {
   createStarryNight,
   common,
-} from "https://esm.sh/@wooorm/starry-night@1";
-import { toHtml } from "https://esm.sh/hast-util-to-html@8";
+} from "https://esm.sh/@wooorm/starry-night@1.7.0";
+import { toHtml } from "https://esm.sh/hast-util-to-html@8.0.4";
 
 // NOTE: "all" is a LOT of code. May want to tone this down if it causes issues.
 const starryNight = await createStarryNight(common);
