--- src/deps/types.ts
@@ -1,1 +1,0 @@
-export type tNav = { name: string; href: string }[];
