+++ src/routes/about.tsx
@@ -1,15 +1,15 @@
 import Layout from "@/components/layout/Layout.tsx";
 import { Handlers, PageProps } from "$fresh/server.ts";
 import Markdown from "@/components/article/Markdown.tsx";
-import { DataPath } from "@/deps/paths.ts";
-import { join } from "$std/path/mod.ts";
 import Contact from "@/components/layout/Contact.tsx";
 import Article from "@/components/article/Article.tsx";
+import { CDN_URL } from "@/deps/paths.ts";
 
 export const handler: Handlers<string> = {
   async GET(_req, ctx) {
-    const data = await Deno.readTextFile(join(DataPath, "about.md"));
-    return ctx.render(data);
+    const articleResp = await fetch(CDN_URL + "/about/index.md");
+    const article = await articleResp.text();
+    return ctx.render(article);
   },
 };
