+++ src/routes/index.tsx
@@ -1,9 +1,9 @@
 import Layout from "@/components/layout/Layout.tsx";
 import { Handlers, PageProps } from "$fresh/server.ts";
-import { getPostMetadata, PostMetadata } from "@/deps/posts.ts";
+import { getPostMetadata, PostMetadata } from "@/cdn/posts.ts";
 import PostCard from "@/components/article/PostCard.tsx";
 import { BorderColor, OutboundLink, TwClass } from "@/deps/styles.ts";
-import { getGfxModuleMetadata, GfxModuleMetadata } from "@/deps/gfx-module.ts";
+import { getGfxModuleMetadata, GfxModuleMetadata } from "@/cdn/gfx-module.ts";
 import GfxCard from "@/components/gfx/GfxCard.tsx";
 import Article from "@/components/article/Article.tsx";
 import ArticleBlurb from "@/components/article/ArticleBlurb.tsx";
@@ -15,6 +15,7 @@ type Props = {
 
 export const handler: Handlers<Props> = {
   async GET(_req, ctx) {
+    _req.method;
     const posts = await getPostMetadata(3);
     const gallery = await getGfxModuleMetadata(3);
     return ctx.render({ posts, gallery });
